+++ lib/debug/Makefile
@@ -6,7 +6,7 @@
 #    By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/31 21:09:57 by kamitsui          #+#    #+#              #
-#    Updated: 2024/12/17 03:16:40 by kamitsui         ###   ########.fr        #
+#    Updated: 2024/12/17 19:18:27 by kamitsui         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -48,6 +48,7 @@ endif
 SRCS = \
 	   put_visited.c \
 	   debug_parse_fc.c \
+	   debug_parse_map_fail.c \
 	   debug_map_data.c \
 	   debug_player.c \
 	   debug_ray.c \
