+++ lib/debug/srcs/debug_parse_map_fail.c
@@ -1,33 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   find_next_element.c                                :+:      :+:    :+:   */
+/*   debug_parse_map_fail.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/15 00:41:18 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/12/17 19:14:03 by kamitsui         ###   ########.fr       */
+/*   Created: 2024/12/17 19:17:43 by kamitsui          #+#    #+#             */
+/*   Updated: 2024/12/17 19:17:54 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-/**
- * @brief Find next line
- *
- * @param line
- *
- * @return exist next line ( NULL : next line is empty )
- */
-const char	*find_next_element(const char *line)
+void	debug_parse_map_fail(int fd, int i)
 {
-	while (line != NULL)
-	{
-		line = find_next_line(line);
-		if (line == NULL)
-			return (NULL);
-		if (*line != '\n')
-			break ;
-	}
-	return (line);
+	if (IS_DEBUG != true)
+		return ;
+	ft_dprintf(fd, "fail : parse_map func[%d]\n", i);
 }
